Added pundit with politics

+++ app/controllers/tags_controller.rb
@@ -4,12 +4,14 @@
 class TagsController < ApplicationController
   def create
     @tag = Tag.new tag_params
+    authorize @tag
     flash[:warning] = @tag.errors.full_messages[0] unless @tag.save
     redirect_back fallback_location: root_path
   end
 
   def destroy
     @tag = Tag.find(params[:id])
+    authorize @tag
     if @tag.destroy
       flash[:success] = 'Tag deleted'
     else
